ecommerce/requirements.txt
aioredis==1.2.0
argon2-cffi==19.1.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.2
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
daphne==2.2.5
dj-database-url==0.5.0
Django==2.1.7
django-bootstrap3==11.0.0
django-braces==1.13.0
Faker==1.0.4
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
misaka==2.1.1
msgpack==0.6.1
Pillow==5.4.1
psycopg2==2.8.2
pycparser==2.19
PyHamcrest==1.9.0
pypiwin32==223
python-dateutil==2.8.0
pytz==2018.9
pywin32==224
requests==2.21.0
six==1.12.0
stripe==2.24.1
text-unidecode==1.2
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.2
wincertstore==0.2
zope.interface==4.6.0
